backend/requirements.txt
Flask==2.1.1
requests==2.26.0
pillow==9.4.0
numpy==1.24.2

timm==0.6.13
opencv-python==4.7.0.72
torch==2.0.0
Flask-Cors==3.0.10
openai==0.27.2
gunicorn==20.1.0